renamed query processor
